Move Position into its own values module
