+++ components/style/values/computed/mod.rs
@@ -14,6 +14,7 @@ pub use cssparser::Color as CSSColor;
 pub use super::specified::{Angle, BorderStyle, Time, UrlExtraData};
 
 pub mod basic_shape;
+pub mod position;
 
 pub struct Context<'a> {
     pub is_root_element: bool,
